+++ src/ts/util/editorCommenEvent.ts
@@ -166,25 +166,49 @@ export const hotkeyEvent = (vditor: IVditor, editorElement: HTMLElement) => {
         }
 
         // delete
-        if (!event.metaKey && !event.ctrlKey && !event.shiftKey && event.keyCode === 8
-            && vditor.currentMode === "markdown") {
-            const position = getSelectPosition(editorElement);
-            if (position.start !== position.end) {
-                insertText(vditor, "", "", true);
+        if (!event.metaKey && !event.ctrlKey && !event.shiftKey && event.keyCode === 8) {
+            if (vditor.currentMode === "markdown") {
+                const position = getSelectPosition(editorElement);
+                if (position.start !== position.end) {
+                    insertText(vditor, "", "", true);
+                } else {
+                    // delete emoji
+                    const text = getText(vditor);
+                    const emojiMatch = text.substring(0, position.start).match(/([\u{1F300}-\u{1F5FF}][\u{2000}-\u{206F}][\u{2700}-\u{27BF}]|([\u{1F900}-\u{1F9FF}]|[\u{1F300}-\u{1F5FF}]|[\u{1F680}-\u{1F6FF}]|[\u{1F600}-\u{1F64F}])[\u{2000}-\u{206F}][\u{2600}-\u{26FF}]|[\u{1F300}-\u{1F5FF}]|[\u{1F100}-\u{1F1FF}]|[\u{1F600}-\u{1F64F}]|[\u{1F680}-\u{1F6FF}]|[\u{1F200}-\u{1F2FF}]|[\u{1F900}-\u{1F9FF}]|[\u{1F000}-\u{1F02F}]|[\u{FE00}-\u{FE0F}]|[\u{1F0A0}-\u{1F0FF}]|[\u{0000}-\u{007F}][\u{20D0}-\u{20FF}]|[\u{0000}-\u{007F}][\u{FE00}-\u{FE0F}][\u{20D0}-\u{20FF}])$/u);
+                    const deleteChar = emojiMatch ? emojiMatch[0].length : 1;
+                    formatRender(vditor,
+                        text.substring(0, position.start - deleteChar) + text.substring(position.start),
+                        {
+                            end: position.start - deleteChar,
+                            start: position.start - deleteChar,
+                        });
+                }
+                event.preventDefault();
+                event.stopPropagation();
             } else {
-                // delete emoji
-                const text = getText(vditor);
-                const emojiMatch = text.substring(0, position.start).match(/([\u{1F300}-\u{1F5FF}][\u{2000}-\u{206F}][\u{2700}-\u{27BF}]|([\u{1F900}-\u{1F9FF}]|[\u{1F300}-\u{1F5FF}]|[\u{1F680}-\u{1F6FF}]|[\u{1F600}-\u{1F64F}])[\u{2000}-\u{206F}][\u{2600}-\u{26FF}]|[\u{1F300}-\u{1F5FF}]|[\u{1F100}-\u{1F1FF}]|[\u{1F600}-\u{1F64F}]|[\u{1F680}-\u{1F6FF}]|[\u{1F200}-\u{1F2FF}]|[\u{1F900}-\u{1F9FF}]|[\u{1F000}-\u{1F02F}]|[\u{FE00}-\u{FE0F}]|[\u{1F0A0}-\u{1F0FF}]|[\u{0000}-\u{007F}][\u{20D0}-\u{20FF}]|[\u{0000}-\u{007F}][\u{FE00}-\u{FE0F}][\u{20D0}-\u{20FF}])$/u);
-                const deleteChar = emojiMatch ? emojiMatch[0].length : 1;
-                formatRender(vditor,
-                    text.substring(0, position.start - deleteChar) + text.substring(position.start),
-                    {
-                        end: position.start - deleteChar,
-                        start: position.start - deleteChar,
-                    });
+                const range = getSelection().getRangeAt(0);
+                if (range.startContainer.nodeType === 3 && range.startOffset === 0) {
+                    const blockRenderElement = range.startContainer.parentElement.previousElementSibling as HTMLElement;
+                    if (blockRenderElement && blockRenderElement.classList.contains('vditor-wysiwyg__block')) {
+
+                        (blockRenderElement.lastElementChild as HTMLElement).click()
+                        event.preventDefault();
+                        event.stopPropagation();
+                    }
+                }
+                if (range.startOffset === 0 && range.startContainer.parentElement.tagName === "CODE" &&
+                    range.startContainer.parentElement.parentElement.parentElement.classList
+                        .contains("vditor-wysiwyg__block")) {
+                    // 渲染代码块中在第一个字符前进行删除时，仅删除代码块，其内容不变
+                    const text = document.createTextNode(range.startContainer.parentElement.textContent)
+                    range.setStartBefore(range.startContainer.parentElement.parentElement.parentElement);
+                    range.insertNode(text);
+                    range.collapse(true);
+                    (vditor.wysiwyg.popover.firstElementChild as HTMLElement).click();
+                    event.preventDefault();
+                    event.stopPropagation();
+                }
             }
-            event.preventDefault();
-            event.stopPropagation();
             return;
         }