+++ src/ts/wysiwyg/processCodeRender.ts
@@ -6,7 +6,7 @@ import {highlightRender} from "../markdown/highlightRender";
 import {mathRenderByLute} from "../markdown/mathRenderByLute";
 import {mermaidRender} from "../markdown/mermaidRender";
 
-// code block, math, math-inline, abc, html, chart, mermaid
+// html, math, math-inline, code block, abc, chart, mermaid
 export const processCodeRender = (blockElement: HTMLElement, vditor: IVditor) => {
     const blockType = blockElement.getAttribute("data-type");
     if (!blockType) {
@@ -18,26 +18,28 @@ export const processCodeRender = (blockElement: HTMLElement, vditor: IVditor) =>
         blockElement.insertAdjacentHTML("beforeend", `<${tagName} class="vditor-wysiwyg__preview"></${tagName}>`);
         previewPanel = blockElement.querySelector(".vditor-wysiwyg__preview");
         previewPanel.setAttribute("data-render", "false");
-        previewPanel.addEventListener("click", (event) => {
+        const showCode = () => {
             const range = preElement.ownerDocument.createRange();
             if (preElement.getAttribute("data-type") === "math-inline") {
                 preElement.style.display = "inline";
                 range.selectNodeContents(preElement);
             } else {
                 preElement.style.display = "block";
-                range.selectNodeContents(preElement.querySelector("code"));
+                range.setStart(preElement.firstElementChild.firstChild, 0);
             }
             range.collapse(true);
             setSelectionFocus(range);
+        };
+        previewPanel.addEventListener("click", () => {
+            showCode();
         });
     }
 
     const preElement = previewPanel.previousElementSibling as HTMLElement;
-
+    const innerHTML = decodeURIComponent((blockElement.querySelector("code").getAttribute("data-code")) || '')
     if (blockType === "code-block") {
         const language = preElement.querySelector("code").className.replace("language-", "");
-        previewPanel.innerHTML =
-            `<pre>${blockElement.firstElementChild.innerHTML}</pre>`;
+        previewPanel.innerHTML = `<pre><code>${innerHTML}</code></pre>`;
         if (language === "abc") {
             abcRender(previewPanel, vditor.options.cdn);
         } else if (language === "mermaid") {
@@ -50,11 +52,9 @@ export const processCodeRender = (blockElement: HTMLElement, vditor: IVditor) =>
             codeRender(previewPanel, vditor.options.lang);
         }
     } else if (blockType.indexOf("html") > -1) {
-        previewPanel.innerHTML =
-            decodeURIComponent(blockElement.querySelector("code").getAttribute("data-code"));
+        previewPanel.innerHTML = innerHTML;
     } else if (blockType.indexOf("math") > -1) {
-        previewPanel.innerHTML =
-            `<${tagName} class="vditor-math">${blockElement.firstChild.textContent}</${tagName}>`;
+        previewPanel.innerHTML = `<${tagName} class="vditor-math">${innerHTML}</${tagName}>`;
         mathRenderByLute(previewPanel, vditor.options.cdn);
     }
 };